Split VoiceSessionSetup into core strategy and web implementation
Break the eager import chain that caused livekit-client to load at
module scope when importing @elevenlabs/client/internal in React Native.

- VoiceSessionStrategy.ts: types + strategy management (zero web imports)
- webSessionSetup.ts: web implementation with static imports
- index.ts registers the web strategy as default via setSetupStrategy()
- New @elevenlabs/client/web entry point for platform code
- RN package uses setup.ts side-effect module to ensure registerGlobals()
  runs before livekit-client loads

Fixes #661
