+++ packages/client/package.json
@@ -10,7 +10,8 @@
       "unpkg": "./dist/lib.iife.js",
       "default": "./dist/index.js"
     },
-    "./internal": "./dist/internal.js"
+    "./internal": "./dist/internal.js",
+    "./web": "./dist/platform/webSessionSetup.js"
   },
   "scripts": {
     "generate-version": "printf \"// This file is auto-generated during build\\nexport const PACKAGE_VERSION = \\\"%s\\\";\\n\" \"$npm_package_version\" > src/version.ts",
