+++ packages/client/src/internal.ts
@@ -7,14 +7,11 @@ export {
 } from "./utils/location.js";
 export { sourceInfo, setSourceInfo } from "./sourceInfo.js";
 export type { SourceInfo } from "./sourceInfo.js";
-export {
-  setSetupStrategy,
-  webSessionSetup,
-} from "./platform/VoiceSessionSetup.js";
+export { setSetupStrategy } from "./platform/VoiceSessionStrategy.js";
 export type {
   VoiceSessionSetupStrategy,
   VoiceSessionSetupResult,
-} from "./platform/VoiceSessionSetup.js";
+} from "./platform/VoiceSessionStrategy.js";
 export {
   MIN_VOICE_FREQUENCY,
   MAX_VOICE_FREQUENCY,
